Moves coil code into new library
