+++ auth/ui/src/main/java/com/google/android/horologist/auth/ui/googlesignin/mapper/AccountUiModelMapper.kt
@@ -18,7 +18,7 @@ package com.google.android.horologist.auth.ui.googlesignin.mapper
 
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.horologist.auth.composables.model.AccountUiModel
-import com.google.android.horologist.compose.material.CoilPaintable
+import com.google.android.horologist.compose.material.coil.CoilPaintable
 
 /**
  * Functions to map models from Google Sign In into a [AccountUiModel].
