+++ compose-material-coil/README.md
@@ -1,0 +1,17 @@
+# Compose Material library - Coil support
+
+[![Maven Central](https://img.shields.io/maven-central/v/com.google.android.horologist/horologist-compose-material-coil)](https://search.maven.org/search?q=g:com.google.android.horologist)
+
+For more information, visit the documentation: https://google.github.io/horologist/compose-material-coil
+
+## Download
+
+```groovy
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+    implementation "com.google.android.horologist:horologist-compose-material-coil:<version>"
+}
+```
