+++ compose-material-coil/api/current.api
@@ -1,0 +1,14 @@
+// Signature format: 4.0
+package com.google.android.horologist.compose.material.coil {
+
+  public final class ArtworkColorKt {
+    method @androidx.compose.runtime.Composable @com.google.android.horologist.annotations.ExperimentalHorologistApi public static androidx.compose.runtime.State<androidx.compose.ui.graphics.Color> rememberArtworkColor(Object? artworkUri, optional long defaultColor);
+  }
+
+  @androidx.compose.runtime.Stable public final class CoilPaintable implements com.google.android.horologist.compose.material.Paintable {
+    ctor public CoilPaintable(Object? model, optional androidx.compose.ui.graphics.painter.Painter? placeholder);
+    method @androidx.compose.runtime.Composable public androidx.compose.ui.graphics.painter.Painter rememberPainter();
+  }
+
+}
+
