+++ compose-material-coil/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=horologist-compose-material-coil
+POM_NAME=Horologist Compose Material Coil library
+POM_PACKAGING=aar
