+++ compose-material-coil/src/main/java/com/google/android/horologist/compose/material/coil/CoilPaintable.kt
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.horologist.compose.material.coil
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Stable
+import androidx.compose.ui.graphics.painter.Painter
+import coil.compose.rememberAsyncImagePainter
+import com.google.android.horologist.compose.material.Paintable
+
+/** A wrapper around a coil-compatible model that can be represented as a [Painter]. */
+@Stable
+public class CoilPaintable(private val model: Any?, private val placeholder: Painter? = null) :
+    Paintable {
+
+        @Composable
+        override fun rememberPainter(): Painter = rememberAsyncImagePainter(
+            model = model,
+            placeholder = placeholder,
+        )
+    }
