+++ compose-material-coil/src/test/snapshots/images/com.google.android.horologist.compose.material_ChipTest_disabledWithIconPlaceholder.png
