+++ compose-material-coil/src/test/snapshots/images/com.google.android.horologist.compose.material_OutlinedCompactChipTest_withPlaceholderIcon.png
