+++ compose-material/src/test/java/com/google/android/horologist/compose/material/CompactChipTest.kt
@@ -20,16 +20,9 @@ import android.R
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.DirectionsBike
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Image
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.unit.LayoutDirection
-import androidx.wear.compose.material.MaterialTheme
 import com.google.accompanist.testharness.TestHarness
 import com.google.android.horologist.compose.material.ImageVectorPaintable.Companion.asPaintable
-import com.google.android.horologist.compose.material.util.rememberVectorPainter
-import com.google.android.horologist.compose.tools.coil.FakeImageLoader
 import com.google.android.horologist.screenshots.ScreenshotBaseTest
 import org.junit.Test
 
@@ -109,54 +102,6 @@ class CompactChipTest : ScreenshotBaseTest() {
                 onClick = { },
                 icon = DrawableResPaintable(R.drawable.ic_delete),
             )
-        }
-    }
-
-    @Test
-    fun withPlaceholderIcon() {
-        screenshotTestRule.setContent(
-            isComponent = true,
-            takeScreenshot = true,
-        ) {
-            // In inspection mode will jump to placeholder
-            CompositionLocalProvider(LocalInspectionMode.provides(true)) {
-                CompactChip(
-                    label = "Primary label",
-                    onClick = { },
-                    icon = CoilPaintable(
-                        "iconUri",
-                        placeholder = rememberVectorPainter(
-                            image = Icons.Filled.Image,
-                            tintColor = Color.Black,
-                        ),
-                    ),
-                )
-            }
-        }
-    }
-
-    @Test
-    fun disabledWithIconPlaceholder() {
-        screenshotTestRule.setContent(
-            isComponent = true,
-            takeScreenshot = true,
-            fakeImageLoader = FakeImageLoader.Never,
-        ) {
-            // In inspection mode will jump to placeholder
-            CompositionLocalProvider(LocalInspectionMode.provides(true)) {
-                CompactChip(
-                    label = "Primary label",
-                    onClick = { },
-                    icon = CoilPaintable(
-                        "iconUri",
-                        placeholder = rememberVectorPainter(
-                            image = Icons.Default.Image,
-                            tintColor = MaterialTheme.colors.primary,
-                        ),
-                    ),
-                    enabled = false,
-                )
-            }
         }
     }
