+++ media/sample/build.gradle.kts
@@ -154,6 +154,7 @@ dependencies {
     implementation(projects.composables)
     implementation(projects.composeLayout)
     implementation(projects.composeMaterial)
+    implementation(projects.composeMaterialCoil)
     implementation(projects.media.core)
     implementation(projects.media.backendMedia3)
     implementation(projects.media.data)
