+++ settings.gradle.kts
@@ -40,6 +40,7 @@ include(":auth:sample:wear")
 include(":auth:ui")
 include(":compose-layout")
 include(":compose-material")
+include(":compose-material-coil")
 include(":compose-tools")
 include(":composables")
 include(":datalayer:core")
